added trainer functional

+++ src/main/java/com/ivan/controller/AthleteController.java
@@ -2,15 +2,18 @@ package com.ivan.controller;
 
 import com.ivan.exception.NotValidArgumentException;
 import com.ivan.model.Athlete;
+import com.ivan.model.Audit;
 import com.ivan.model.Training;
 import com.ivan.model.TrainingType;
 import com.ivan.service.AthleteService;
+import com.ivan.service.AuditService;
 import com.ivan.service.TrainingService;
 import com.ivan.service.TrainingTypeService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
 import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
 import java.util.Arrays;
 import java.util.List;
 
@@ -19,6 +22,7 @@ import java.util.List;
 public class AthleteController {
 
     private final AthleteService athleteService;
+    private final AuditService auditService;
     private final TrainingService trainingService;
     private final TrainingTypeService trainingTypeService;
 
@@ -31,12 +35,20 @@ public class AthleteController {
             throw new NotValidArgumentException("Enter a number greater than 0. letters cannot be used!");
         }
 
+        if (!isValidDate(date)) {
+            throw new NotValidArgumentException("Incorrect date format. it should match yyyy-MM-dd!");
+        }
+
         trainingService.addTraining(athleteId, trainingType, Integer.parseInt(setsAmount), date);
     }
 
     public void editTrainingByAthleteIdAndTrainingDate(Long athleteId, LocalDate date, String trainingType, String setsAmount) {
         if (!isValidNum(setsAmount)) {
             throw new NotValidArgumentException("Enter a number greater than 0. letters cannot be used!");
+        }
+
+        if (!isValidDate(date)) {
+            throw new NotValidArgumentException("Incorrect date format. it should match yyyy-MM-dd!");
         }
 
         trainingService.updateTraining(athleteId, date, trainingType, setsAmount);
@@ -47,14 +59,18 @@ public class AthleteController {
     }
 
     public List<Training> showHistoryTrainingsBySetsAmount(Long athleteId) {
-        return trainingService.getTrainingsBySetsAmount(athleteId);
+        return trainingService.getTrainingsSortedBySetsAmount(athleteId);
     }
 
     public void deleteTraining(Long athleteId, LocalDate date) {
+        if (!isValidDate(date)) {
+            throw new NotValidArgumentException("Incorrect date format. it should match yyyy-MM-dd!");
+        }
+
         trainingService.deleteTraining(athleteId, date);
     }
 
-    public List<Athlete> showAllUser() {
+    public List<Athlete> showAllClients() {
         return athleteService.showAllAthlete();
     }
 
@@ -62,8 +78,28 @@ public class AthleteController {
         trainingTypeService.save(trainingType);
     }
 
+    public void deleteTrainingType(String trainingTypeName) {
+        trainingTypeService.delete(trainingTypeName);
+    }
+
+    public List<Audit> showAuditsByAthleteId(String athlete) {
+        return auditService.getAllAuditsByAthleteId(athlete);
+    }
+
     private boolean isValidNum(String... inputs) {
-        return Arrays.stream(inputs)
-                .allMatch(input -> input.matches("\\d+"));
+        for (String input : inputs) {
+            if (!input.matches("\\d+")) {
+                return false;
+            }
+        }
+        return true;
+    }
+
+    private boolean isValidDate(LocalDate date) {
+        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
+
+        String formattedDate = date.format(formatter);
+
+        return formattedDate.equals(date.toString());
     }
 }
